addition of necessary UI and small refactor for readability
