+++ ts/features/pn/navigation/navigator.tsx
@@ -2,7 +2,7 @@ import { createStackNavigator } from "@react-navigation/stack";
 import { MessageAttachmentScreen } from "../screens/MessageAttachmentScreen";
 import { MessageDetailsScreen } from "../screens/MessageDetailsScreen";
 import { PaidPaymentScreen } from "../screens/PaidPaymentScreen";
-import { PNActivationBannerFlowScreen } from "../screens/PnReminderBannerFlow";
+import { PNActivationBannerFlowScreen } from "../reminderBanner/screens/PnReminderBannerFlow";
 import { PnParamsList } from "./params";
 import PN_ROUTES from "./routes";
